Set Cache-Control explicitly per page via sendPage instead of a router-wide middleware
