+++ packages/frontend/src/pages/ServerPageRouter.ts
@@ -1,9 +1,5 @@
 import express from 'express'
 import { env } from '~/env'
-import {
-  ClearPageCacheMiddleware,
-  PageCacheMiddleware,
-} from '~/server/middlewares/PageCacheMiddleware'
 import { FrontendInMemoryCache } from '~/utils/FrontendInMemoryCache'
 import type { RenderFunction } from '../ssr/types'
 import type { Manifest } from '../utils/Manifest'
@@ -47,10 +43,6 @@ export function createServerPageRouter(
     next()
   })
 
-  // Cloudflare edge-caches HTML only when the origin sends Cache-Control.
-  // Routes that must not be cached (e.g. "/") override it later in the chain.
-  router.use('/', PageCacheMiddleware())
-
   if (!env.CLIENT_SIDE_HOME_PAGE) {
     // Temporary redirect so browsers drop the previously cached 301 before
     // "/" starts serving the home page. no-cache (not no-store) so the
@@ -93,10 +85,6 @@ export function createServerPageRouter(
       router.use('/', subRouter)
     }
   }
-
-  // Anything reaching here was not a page (e.g. /api/*, /health, 404s) and
-  // must not be edge-cached.
-  router.use('/', ClearPageCacheMiddleware())
 
   return router
 }
