+++ packages/frontend/src/pages/brand-kit/BrandKitRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getBrandKitData } from './getBrandKitData'
@@ -12,7 +13,7 @@ export function createBrandKitRouter(
   router.get('/brand-kit', async (req, res) => {
     const data = await getBrandKitData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
