+++ packages/frontend/src/pages/da-risk-framework/DaRiskFrameworkRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { getDaRiskFrameworkData } from './getDaRiskFrameworkData'
@@ -12,7 +13,7 @@ export function createDaRiskFrameworkRouter(
   router.get('/da-risk-framework', async (req, res) => {
     const data = getDaRiskFrameworkData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
