+++ packages/frontend/src/pages/data-availability/DataAvailabilityRouter.ts
@@ -1,6 +1,7 @@
 import type { InMemoryCache } from '@l2beat/shared-pure'
 import { v } from '@l2beat/validate'
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { validateRoute } from '~/utils/validateRoute'
@@ -32,7 +33,7 @@ export function createDataAvailabilityRouter(
       () => getDataAvailabilitySummaryData(manifest, req.originalUrl),
     )
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/data-availability/risk', async (req, res) => {
@@ -45,7 +46,7 @@ export function createDataAvailabilityRouter(
       () => getDataAvailabilityRiskData(manifest, req.originalUrl),
     )
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/data-availability/throughput', async (req, res) => {
@@ -58,7 +59,7 @@ export function createDataAvailabilityRouter(
       () => getDataAvailabilityThroughputData(manifest, req.originalUrl),
     )
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/data-availability/liveness', async (req, res) => {
@@ -71,7 +72,7 @@ export function createDataAvailabilityRouter(
       () => getDataAvailabilityLivenessData(manifest, req.originalUrl),
     )
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/data-availability/archived', async (req, res) => {
@@ -84,7 +85,7 @@ export function createDataAvailabilityRouter(
       () => getDataAvailabilityArchivedData(manifest, req.originalUrl),
     )
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get(
@@ -112,7 +113,7 @@ export function createDataAvailabilityRouter(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
