+++ packages/frontend/src/pages/dev/DevRouter.ts
@@ -1,5 +1,6 @@
 import express from 'express'
 import { env } from '~/env'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getIconPreviewData } from './icons/getIconPreviewData'
@@ -14,7 +15,7 @@ export function createDevRouter(manifest: Manifest, render: RenderFunction) {
   router.get('/dev/icons', async (req, res) => {
     const data = await getIconPreviewData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
