+++ packages/frontend/src/pages/donate/DonateRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getDonateData } from './getDonateData'
@@ -9,7 +10,7 @@ export function createDonateRouter(manifest: Manifest, render: RenderFunction) {
   router.get('/donate', async (req, res) => {
     const data = await getDonateData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
