+++ packages/frontend/src/pages/faq/FaqRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getFaqData } from './getFaqData'
@@ -9,7 +10,7 @@ export function createFaqRouter(manifest: Manifest, render: RenderFunction) {
   router.get('/faq', async (req, res) => {
     const data = await getFaqData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
