+++ packages/frontend/src/pages/glossary/GlossaryRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getGlossaryData } from './getGlossaryData'
@@ -12,7 +13,7 @@ export function createGlossaryRouter(
   router.get('/glossary', async (req, res) => {
     const data = await getGlossaryData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
