+++ packages/frontend/src/pages/interop/InteropRouter.ts
@@ -2,6 +2,7 @@ import type { InMemoryCache } from '@l2beat/shared-pure'
 import { v } from '@l2beat/validate'
 import express from 'express'
 import { ps } from '~/server/projects'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { validateRoute } from '~/utils/validateRoute'
@@ -48,7 +49,7 @@ export function createInteropRouter(
     async (req, res) => {
       const data = await getInteropSummaryData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
@@ -60,7 +61,7 @@ export function createInteropRouter(
     async (req, res) => {
       const data = await getInteropNonMintingData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
@@ -72,7 +73,7 @@ export function createInteropRouter(
     async (req, res) => {
       const data = await getInteropLockAndMintData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
@@ -84,20 +85,20 @@ export function createInteropRouter(
     async (req, res) => {
       const data = await getInteropBurnAndMintData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
   router.get('/interop/token-frameworks', async (req, res) => {
     const data = await getInteropTokenFrameworksData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/interop/intent-bridges', async (req, res) => {
     const data = await getInteropIntentBridgesData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get(
@@ -125,7 +126,7 @@ export function createInteropRouter(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
@@ -164,7 +165,7 @@ export function createInteropRouter(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
