+++ packages/frontend/src/pages/layer2s/L2Router.ts
@@ -2,6 +2,7 @@ import type { InMemoryCache } from '@l2beat/shared-pure'
 import { v } from '@l2beat/validate'
 import express from 'express'
 import { env } from '~/env'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { validateRoute } from '~/utils/validateRoute'
@@ -33,37 +34,37 @@ export function createL2Router(
   router.get('/layer2s/summary', async (req, res) => {
     const data = await getL2SummaryData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/activity', async (req, res) => {
     const data = await getL2ActivityData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/risk', async (req, res) => {
     const data = await getL2RiskData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/risk/state-validation', async (req, res) => {
     const data = await getL2RiskStateValidationData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/risk/data-availability', async (req, res) => {
     const data = await getL2RiskDataAvailabilityData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/risk/sequencing', async (req, res) => {
     const data = await getL2RiskSequencingData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get(
@@ -78,20 +79,20 @@ export function createL2Router(
     async (req, res) => {
       const data = await getL2TvsData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
   router.get('/layer2s/tvs/breakdown', async (req, res) => {
     const data = await getL2TvsBreakdownData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get('/layer2s/liveness', async (req, res) => {
     const data = await getL2LivenessData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   if (env.CLIENT_SIDE_COMPARE_PROJECTS) {
@@ -100,7 +101,7 @@ export function createL2Router(
     router.get('/layer2s/compare', async (req, res) => {
       const data = await getL2CompareData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     })
   }
 
@@ -114,14 +115,14 @@ export function createL2Router(
     async (req, res) => {
       const data = await getL2CostsData(req, manifest, cache)
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
   router.get('/layer2s/archived', async (req, res) => {
     const data = await getL2ArchivedData(req, manifest, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get(
@@ -137,7 +138,7 @@ export function createL2Router(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
 
@@ -165,7 +166,7 @@ export function createL2Router(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
