+++ packages/frontend/src/pages/multisig-report/MutlisigReportRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '~/utils/Manifest'
 import { getMultisigReportData } from './getMultisigReportData'
@@ -12,7 +13,7 @@ export function createMultisigReportRouter(
   router.get('/multisig-report', async (req, res) => {
     const data = await getMultisigReportData(manifest, req.originalUrl)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
