+++ packages/frontend/src/pages/stages/StagesRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getStagesData } from './getStagesData'
@@ -13,7 +14,7 @@ export function createStagesRouter(manifest: Manifest, render: RenderFunction) {
       return
     }
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
