+++ packages/frontend/src/pages/terms-of-service/TermsOfServiceRouter.ts
@@ -1,4 +1,5 @@
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import type { Manifest } from '../../utils/Manifest'
 import { getTermsOfServiceData } from './getTermsOfServiceData'
@@ -16,7 +17,7 @@ export function createTermsOfServiceRouter(
       return
     }
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   return router
