+++ packages/frontend/src/pages/zk-catalog/ZkCatalogRouter.ts
@@ -1,6 +1,7 @@
 import type { InMemoryCache } from '@l2beat/shared-pure'
 import { v } from '@l2beat/validate'
 import express from 'express'
+import { sendPage } from '~/server/utils/sendPage'
 import type { RenderFunction } from '~/ssr/types'
 import { validateRoute } from '~/utils/validateRoute'
 import type { Manifest } from '../../utils/Manifest'
@@ -17,7 +18,7 @@ export function createZkCatalogRouter(
   router.get('/zk-catalog', async (req, res) => {
     const data = await getZkCatalogData(manifest, req.originalUrl, cache)
     const html = await render(data, req.originalUrl)
-    res.status(200).send(html)
+    sendPage(res, html)
   })
 
   router.get(
@@ -40,7 +41,7 @@ export function createZkCatalogRouter(
         return
       }
       const html = await render(data, req.originalUrl)
-      res.status(200).send(html)
+      sendPage(res, html)
     },
   )
